src/ingestion/document_processing.py
from azure.ai.documentintelligence import DocumentIntelligenceClient
from azure.core.credentials import AzureKeyCredential

def analyze_pdf(endpoint, key, blob_url):
    client = DocumentIntelligenceClient(endpoint=endpoint, credential=AzureKeyCredential(key))
    poller = client.begin_analyze_document("prebuilt-read", document=blob_url)
    result = poller.result()
    return result
