requirements.txt

langchain
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters
azure-storage-blob
azure-ai-documentintelligence
azure-search-documents
openai
